cambios1

+++ src/main/java/com/intersoluciones/repository/TipoDocumentoRepository.java
@@ -4,7 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.intersoluciones.entities.TipoDocumento;
 
-public interface TipoDocumentoRepository extends JpaRepository<TipoDocumento,Integer> {
-
+public interface TipoDocumentoRepository extends JpaRepository<TipoDocumento,Long> {
 
 }
